feat: add a dataset_reader for custom typos datasets
